backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.0
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.0.0
django-ninja==1.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.8
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
requests-mock==1.12.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
pydantic[email]
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.0.0
django-ninja==1.3.0
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.2.0
gunicorn==20.1.0
idna==3.8
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
requests-mock==1.12.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
annotated-types==0.7.0
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.22.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-ninja==1.3.0
django-storages==1.14.4
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.2.0
gunicorn==20.1.0
idna==3.8
isodate==0.6.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
redis==4.3.4
django-redis==5.2.0